Control de errores

+++ application/controller.py
@@ -1,14 +1,27 @@
+from urllib.error import URLError
 from application.model import event, attendee
 from application.model.category import Category
 from application.services.attendee_service import AttendeeService
 from application.services.category_service import CategoryService
 from application.services.event_service import EventService
 from application.services.place_service import PlaceService
+from werkzeug.exceptions import BadRequest
 
 __author__ = 'Dani Meana'
 
 import flask
 from application import app
+
+
+@app.errorhandler(Exception)
+def all_exception_handler(error):
+    return _response_error(code="dit_err", message="Server critical error. Please, contact with administrators",
+                           status=500)
+
+
+@app.errorhandler(404)
+def resource_not_found(error):
+    return _response_error(code="dit_404", message="Resource not found", status=404)
 
 
 @app.route("/createTables")
@@ -21,13 +34,15 @@ def create():
 
     for category in data.categories:
         CategoryService.save(
-            Category(id_category=category["id"], color=category["color"], image=category["image"]))
+            Category(id_category=category['id'],
+                     color=category['color'],
+                     image=category['image']))
 
     for i in range(0, 10):
         for event_dict in data.events:
             EventService.save(event.from_dict(event_dict))
 
-    return "<h1> Tablas creadas</h1>"
+    return "<h1>Tablas creadas</h1>"
 
 
 @app.route("/", methods=["GET"])
@@ -42,36 +57,54 @@ def find_near_events():
     radius = _get_request_arg('radius', None)
     from_id = _get_request_arg('fromId', None)
     elements = _get_request_arg('elements', None)
-    return _collection_to_json(EventService.find_near_events(lat, lng, radius, from_id, elements))
+    try:
+        return _collection_to_json(EventService.find_near_events(lat, lng, radius, from_id, elements))
+    except ValueError:
+        return _response_error(code="dit_fes", message="The arguments are not correct")
 
 
 @app.route("/events", methods=["POST"])
 def save_event():
-    event_json = flask.request.get_json()
-    event_obj = event.from_dict(event_json)
-    EventService.save(event_obj)
-    return _response_ok()
+    try:
+        event_json = flask.request.get_json()
+        event_obj = event.from_dict(event_json)
+        EventService.save(event_obj)
+        return _response_ok(code="dit_se", message="Event saved")
+    except (KeyError, ValueError, BadRequest):
+        return _response_error(code="dit_se", message="The event to save is wrong")
 
 
 @app.route("/events/<int:event_id>", methods=["GET"])
 def find_event(event_id):
-    return _element_to_json(EventService.get(event_id))
+    event_obj = EventService.get(event_id)
+    if event_obj is None:
+        return _response_error(code="dit_fe", message="Event not found")
+    return _element_to_json(event_obj)
 
 
 @app.route("/events/<int:event_id>", methods=["DELETE"])
 def delete_event(event_id):
     event_obj = EventService.get(event_id)
+    if event_obj is None:
+        return _response_error(code="dit_de", message="Event not found")
     EventService.delete(event_obj)
-    return _response_ok()
+    return _response_ok(code="dit_de", message="Event deleted")
 
 
 @app.route("/events/<int:event_id>", methods=["PUT"])
 def update_event(event_id):
-    event_json = flask.request.get_json()
-    event_obj = event.from_dict(event_json)
-    event_obj.id = event_id
-    EventService.update(event_obj)
-    return _response_ok()
+    try:
+        event_json = flask.request.get_json()
+        event_obj = event.from_dict(event_json)
+    except (KeyError, BadRequest):
+        return _response_error(code="dit_ue", message="Event to update is wrong")
+
+    try:
+        event_obj.id = event_id
+        EventService.update(event_obj)
+        return _response_ok(code="dit_ue", message="Event updated")
+    except (ValueError, AttributeError):
+        return _response_error(code="dit_ue", message="Event to update does not exist")
 
 
 @app.route("/events/<int:event_id>/attendees", methods=["GET"])
@@ -81,24 +114,40 @@ def find_attendees(event_id):
 
 @app.route("/events/<int:event_id>/attendees", methods=["POST"])
 def add_attendee(event_id):
-    attendee_json = flask.request.get_json()
-    attendee_obj = attendee.from_dict(attendee_json)
-    # Check ids coinciden
+    try:
+        attendee_json = flask.request.get_json()
+        attendee_obj = attendee.from_dict(attendee_json)
+    except (KeyError, BadRequest):
+        return _response_error(code="dit_aa", message="Attendee to add to an event is wrong")
+    if event_id is not attendee_obj.event_id:
+        return _response_error(code="dit_aa", message="The event of the attendee is not correct")
+    event_obj = EventService.get(event_id)
+    if event_obj is None:
+        return _response_error(code="dit_aa", message="Event not found")
+    attendee_old = AttendeeService.get(event_id, attendee_obj.user_id)
+    if attendee_old is not None:
+        return _response_ok(code="dit_aa", message="Attendee already exists")
     AttendeeService.save(attendee_obj)
-    return _response_ok()
+    return _response_ok(code="dit_aa", message="Attendee added")
+
 
 @app.route("/events/<int:event_id>/attendees/<string:user_id>", methods=["DELETE"])
 def delete_attendee(event_id, user_id):
     attendee_obj = AttendeeService.get(event_id, user_id)
+    if attendee_obj is None:
+        return _response_error(code="dit_da", message="Attendee not found")
     AttendeeService.delete(attendee_obj)
-    return _response_ok()
+    return _response_ok(code="dit_da", message="Attendee deleted")
 
 
 @app.route("/users/<string:user_id>/events", methods=["GET"])
 def find_events_by_user(user_id):
     from_id = _get_request_arg('fromId', None)
     elements = _get_request_arg('elements', None)
-    return _collection_to_json(EventService.find_events_by_user_id(user_id, from_id, elements))
+    try:
+        return _collection_to_json(EventService.find_events_by_user_id(user_id, from_id, elements))
+    except ValueError:
+        return _response_error(code="dit_fesu", message="The arguments are not correct")
 
 
 @app.route("/categories/<int:category_id>/events", methods=["GET"])
@@ -108,13 +157,19 @@ def find_near_events_by_category(category_id):
     radius = _get_request_arg('radius', None)
     from_id = _get_request_arg('fromId', None)
     elements = _get_request_arg('elements', None)
-    return _collection_to_json(
-        EventService.find_near_events_by_category_id(category_id, lat, lng, radius, from_id, elements))
+    try:
+        return _collection_to_json(
+            EventService.find_near_events_by_category_id(category_id, lat, lng, radius, from_id, elements))
+    except ValueError:
+        return _response_error(code="dit_fesc", message="The arguments are not correct")
 
 
 @app.route("/categories")
 def find_categories():
-    return _collection_to_json(CategoryService.get_all())
+    try:
+        return _collection_to_json(CategoryService.get_all())
+    except URLError:
+        return _response_error(code="dit_fc", message="The server is not available", status=500)
 
 
 @app.route("/categories/<int:category_id>/places", methods=["GET"])
@@ -123,7 +178,10 @@ def find_near_places_by_category(category_id):
     lng = _get_request_arg('lng', None)
     radius = _get_request_arg('radius', None)
     elements = _get_request_arg('elements', None)
-    return _collection_to_json(PlaceService.find_near_places_by_category(category_id, lat, lng, radius, elements))
+    try:
+        return _collection_to_json(PlaceService.find_near_places_by_category(category_id, lat, lng, radius, elements))
+    except URLError:
+        return _response_error(code="dit_fpc", message="The server is not available", status=500)
 
 
 def _get_request_arg(arg, default_value):
@@ -139,12 +197,26 @@ def _element_to_json(element):
 
 
 def _response(data):
-    """Create a JSON response."""
     return flask.Response(response=flask.json.dumps(data),
                           status=200,
-                          mimetype="application/json")
+                          mimetype='application/json')
 
 
-def _response_ok():
-    return flask.Response(status=200,
-                          mimetype="application/json")
+def _response_ok(code, message):
+    data = {
+        'code': code,
+        'message': message
+    }
+    return flask.Response(response=flask.json.dumps(data),
+                          status=200,
+                          mimetype='application/json')
+
+
+def _response_error(code, message, status=400):
+    data = {
+        'code': code,
+        'message': message
+    }
+    return flask.Response(response=flask.json.dumps(data),
+                          status=status,
+                          mimetype='application/json')
